requirements.txt
streamlit
supabase
plotly
folium
streamlit-folium
tensorflow
opencv-python
pillow
requests
